Move control of style book to global styles panel.
