+++ packages/edit-site/src/components/sidebar-global-styles-wrapper/style.scss
@@ -29,4 +29,7 @@
 			padding-right: $grid-unit-30;
 		}
 	}
+	.edit-site-sidebar-button {
+		color: $gray-900;
+	}
 }
